+++ h2o-algos/src/main/java/hex/tree/drf/DRF.java
@@ -321,7 +321,7 @@ public class DRF extends SharedTree<hex.tree.drf.DRFModel, hex.tree.drf.DRFModel
       final boolean importance = true;
       @Override public void map( Chunk[] chks ) {
         final Chunk    y       = importance ? chk_resp(chks) : null; // Response
-        final float [] rpred   = importance ? new float [1+_nclass] : null; // Row prediction
+        final double[] rpred   = importance ? new double[1+_nclass] : null; // Row prediction
         final double[] rowdata = importance ? new double[_ncols] : null; // Pre-allocated row data
         final Chunk   oobt  = chk_oobt(chks); // Out-of-bag rows counter over all trees
         // Iterate over all rows
@@ -369,7 +369,7 @@ public class DRF extends SharedTree<hex.tree.drf.DRFModel, hex.tree.drf.DRFModel
                 int actuPred = (int) y.at8(row);
                 if (treePred==actuPred) rightVotes++; // No miss !
               } else { // regression
-                float  treePred = rpred[1];
+                double treePred = rpred[1];
                 double actuPred = y.atd(row);
                 sse += (actuPred-treePred)*(actuPred-treePred);
               }
@@ -483,17 +483,17 @@ public class DRF extends SharedTree<hex.tree.drf.DRFModel, hex.tree.drf.DRFModel
   // Read the 'tree' columns, do model-specific math and put the results in the
   // fs[] array, and return the sum.  Dividing any fs[] element by the sum
   // turns the results into a probability distribution.
-  @Override protected float score1( Chunk chks[], float fs[/*nclass*/], int row ) {
+  @Override protected double score1( Chunk chks[], double fs[/*nclass*/], int row ) {
     double sum = 0;
     if (_nclass > 1) { //classification
       for (int k = 0; k < _nclass; k++)
-        sum += (fs[k+1] = (float) chk_tree(chks, k).atd(row));
+        sum += (fs[k+1] = chk_tree(chks, k).atd(row));
     } else { //regression
       // average per trees voted for this row (only trees which have row in "out-of-bag"
-      sum += (fs[0] = (float) chk_tree(chks, 0).atd(row) / (float)chk_oobt(chks).atd(row) );
+      sum += (fs[0] = chk_tree(chks, 0).atd(row) / chk_oobt(chks).atd(row) );
       fs[1] = 0;
     }
-    return (float)sum;
+    return sum;
   }
