+++ h2o-algos/src/test/java/hex/deeplearning/DeepLearningMLPReference.java
@@ -720,7 +720,7 @@ public class DeepLearningMLPReference {
         //int maxIndex = MaxIndex(yValues); // which cell in yValues has largest value?
 
         // convert to float and do the same tie-breaking as H2O
-        float[] preds = new float[yValues.length+1];
+        double[] preds = new double[yValues.length+1];
         for (int j=0; j<yValues.length; ++j) preds[j+1] = (float)yValues[j];
         preds[0] = hex.genmodel.GenModel.getPrediction(preds, xValues);
